requirements.txt
cachetools==6.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.4.0
cuml-cu12==25.4.0
cupy-cuda12x==13.4.1
cuvs-cu12==25.4.0
cycler==0.12.1
dask==2025.2.0
dask-cuda==25.4.0
dask-cudf-cu12==25.4.0
distributed==2025.2.0
distributed-ucxx-cu12==0.43.0
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
hdbscan==0.8.40
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Levenshtein==0.27.1
libcudf-cu12==25.4.0
libcuml-cu12==25.4.0
libcuvs-cu12==25.4.0
libkvikio-cu12==25.4.0
libraft-cu12==25.4.0
librmm-cu12==25.4.0
libucx-cu12==1.18.1
libucxx-cu12==0.43.0
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numba-cuda==0.4.0
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.12
packaging==25.0
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
polars==1.30.0
psutil==7.0.0
pyarrow==19.0.1
Pygments==2.19.1
pylibcudf-cu12==25.4.0
pylibraft-cu12==25.4.0
pynndescent==0.5.13
pynvjitlink-cu12==0.6.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
raft-dask-cu12==25.4.0
RapidFuzz==3.13.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rmm-cu12==25.4.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
treelite==4.4.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
ucx-py-cu12==0.43.0
ucxx-cu12==0.43.0
umap-learn==0.5.7
urllib3==2.4.0
zict==3.0.0
